feat: Updated router structure to support credentials create
