+++ apps/website/components/templates/protocol/credentials/show/components/credential-card-info.tsx
