+++ apps/website/components/templates/protocol/credentials/show/components/credential-title-and-image.tsx
