+++ apps/website/components/templates/protocol/data-models/show/components/data-model-tabs.tsx
