+++ apps/website/components/templates/protocol/data-models/show/data-model-show.tsx
@@ -1,4 +1,6 @@
 import useTranslation from 'next-translate/useTranslation';
+import { useRouter } from 'next/router';
+import { useEffect } from 'react';
 
 import { PartialDeep } from 'type-fest/source/partial-deep';
 
@@ -6,17 +8,30 @@ import { TOKENS } from '@gateway/theme';
 
 import { Stack, Typography, Button } from '@mui/material';
 
+import { ROUTES } from '../../../../../constants/routes';
 import { DataModel } from '../../../../../services/gateway-protocol/types';
 import InfoTitle from '../../components/info-title';
 import Tags from '../../components/tags';
+import CredentialProtocolCreate from '../../credentials/create/credential-create';
 import DataModelTabs from './components/data-model-tabs';
 
 type Props = {
   dataModel: PartialDeep<DataModel>;
+  isCredentialCreate?: boolean;
 };
 
-export default function DataModelView({ dataModel }: Props) {
+export default function DataModelShow({
+  dataModel,
+  isCredentialCreate = false,
+}: Props) {
   const { t } = useTranslation('protocol');
+  const router = useRouter();
+
+  useEffect(() => {
+    if (isCredentialCreate) {
+      document.body.style.overflow = 'hidden';
+    }
+  }, [isCredentialCreate]);
 
   return (
     <>
@@ -30,11 +45,21 @@ export default function DataModelView({ dataModel }: Props) {
         />
         <Tags tags={dataModel?.tags} />
         <Typography sx={{ mb: 3 }}>{dataModel?.description}</Typography>
-        <Button variant="contained" disabled={true} sx={{ width: '180px' }}>
+        <Button
+          variant="contained"
+          sx={{ width: '180px' }}
+          onClick={() => {
+            router.push({
+              pathname: ROUTES.PROTOCOL_DATAMODEL_CREDENTIAL_CREATE,
+              query: { id: router?.query?.id },
+            });
+          }}
+        >
           {t('data-model.issue-credential-button')}
         </Button>
       </Stack>
       <DataModelTabs dataModel={dataModel} />
+      {isCredentialCreate && <CredentialProtocolCreate />}
     </>
   );
 }
