+++ apps/website/constants/routes.js
@@ -19,8 +19,10 @@ const ROUTES = {
   DAO_NEW: '/dao/new',
   GATE_PROFILE: '/credential/[id]',
   GATE_NEW: '/credential/new',
-  PROTOCOL_CREDENTIAL: '/protocol/credential/[id]',
-  PROTOCOL_DATAMODEL: '/protocol/data-model/[id]',
+  PROTOCOL_CREDENTIAL: '/protocol/credentials/[id]/show',
+  PROTOCOL_DATAMODEL: '/protocol/data-models/[id]/show',
+  PROTOCOL_DATAMODEL_CREDENTIAL_CREATE:
+    '/protocol/data-models/[id]/credentials/create',
 };
 
 module.exports = {
