+++ apps/website/pages/protocol/credentials/[id]/show.tsx
@@ -1,11 +1,11 @@
 import { InferGetServerSidePropsType, GetServerSidePropsContext } from 'next';
 
-import { DashboardTemplate } from '../../../components/templates/dashboard';
+import { DashboardTemplate } from '../../../../components/templates/dashboard';
 import {
-  CredentialProtocolView,
+  CredentialProtocolShow,
   ProtocolTemplate,
-} from '../../../components/templates/protocol';
-import { gatewayProtocolSDK } from '../../../services/gateway-protocol/api';
+} from '../../../../components/templates/protocol';
+import { gatewayProtocolSDK } from '../../../../services/gateway-protocol/api';
 
 type Props = InferGetServerSidePropsType<typeof getServerSideProps>;
 
@@ -20,7 +20,7 @@ export default function ProtocolCredential({ credential }: Props) {
       }}
     >
       <ProtocolTemplate>
-        <CredentialProtocolView credential={credential} />
+        <CredentialProtocolShow credential={credential} />
       </ProtocolTemplate>
     </DashboardTemplate>
   );
