+++ apps/website/pages/protocol/data-models/[id]/show.tsx
@@ -1,0 +1,39 @@
+import { InferGetServerSidePropsType, GetServerSidePropsContext } from 'next';
+
+import { DashboardTemplate } from '../../../../components/templates/dashboard';
+import {
+  ProtocolTemplate,
+  DataModelShow,
+} from '../../../../components/templates/protocol';
+import { gatewayProtocolSDK } from '../../../../services/gateway-protocol/api';
+
+type Props = InferGetServerSidePropsType<typeof getServerSideProps>;
+
+export default function ProtocolDataModel({ dataModel }: Props) {
+  return (
+    <DashboardTemplate
+      containerProps={{
+        sx: {
+          overflow: '',
+        },
+        height: '100%',
+      }}
+    >
+      <ProtocolTemplate>
+        <DataModelShow dataModel={dataModel} />
+      </ProtocolTemplate>
+    </DashboardTemplate>
+  );
+}
+
+export const getServerSideProps = async (ctx: GetServerSidePropsContext) => {
+  const dataModel = await gatewayProtocolSDK.dataModel({
+    id: ctx.query.id as string,
+  });
+
+  return {
+    props: {
+      dataModel: dataModel?.dataModel,
+    },
+  };
+};
